修复bug: 不在数据库内的物品会导致判断语句报错

+++ core.lua
@@ -7,7 +7,7 @@ function filterTrashes( ... )
 	for b=0,4,1 do 
 		for s =1,32,1 do 
 			local id=select(10, GetContainerItemInfo(b,s)) 
-			if fishes[id].delete or trashes[id].delete then 
+			if (fishes[id] and fishes[id].delete) or (trashes[id] and trashes[id].delete) then 
 				PickupContainerItem(b,s)
 				DeleteCursorItem() 
 			end
